Arrange all principal fuunctions in their own JS files. Create session file, which storage the last appendChild to main to keep a fluid usage. Add a delete project button, which delete all the projects inside. Add styles to form and cards showing each todo.
