+++ assets/css/stylesheet.css
@@ -187,3 +187,18 @@ textarea {
 .card:hover {
   filter: brightness(0.8)
 }
+
+.delete-project-btn{
+  border: 2px solid red;
+  color: red;
+  padding: 5px;
+  border-radius: 3px;
+  width: auto;
+  font-size: 1.2em;
+  position: absolute;
+  bottom: 10px;
+  right: 10px;
+}
+.delete-project-btn:hover{
+  cursor: pointer;
+}
